+++ zarrs/examples/async_array_write_read.rs
@@ -9,7 +9,7 @@ async fn async_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
     use futures::StreamExt;
     use std::sync::Arc;
     use zarrs::{
-        array::{DataType, FillValue, ZARR_NAN_F32},
+        array::{DataType, ZARR_NAN_F32},
         array_subset::ArraySubset,
         node::Node,
     };
@@ -55,8 +55,8 @@ async fn async_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
     let array = zarrs::array::ArrayBuilder::new(
         vec![8, 8], // array shape
         DataType::Float32,
-        vec![4, 4].try_into()?, // regular chunk shape
-        FillValue::from(ZARR_NAN_F32),
+        vec![4, 4], // regular chunk shape
+        ZARR_NAN_F32,
     )
     // .bytes_to_bytes_codecs(vec![]) // uncompressed
     .dimension_names(["y", "x"].into())