+++ zarrs/examples/custom_data_type_uint4.rs
@@ -207,7 +207,7 @@ fn main() {
     let array = ArrayBuilder::new(
         vec![6, 1], // array shape
         DataType::Extension(Arc::new(CustomDataTypeUInt4)),
-        vec![5, 1].try_into().unwrap(), // regular chunk shape
+        vec![5, 1], // regular chunk shape
         FillValue::new(fill_value.to_ne_bytes().to_vec()),
     )
     .array_to_array_codecs(vec![